Use StreamWorker interface to allow for KTable implementations
