+++ src/main/java/org/radarcns/stream/empatica/E4TemperatureStream.java
@@ -25,7 +25,7 @@ import org.radarcns.kafka.ObservationKey;
 import org.radarcns.passive.empatica.EmpaticaE4Temperature;
 import org.radarcns.stream.StreamDefinition;
 import org.radarcns.stream.StreamMaster;
-import org.radarcns.stream.StreamWorker;
+import org.radarcns.stream.KStreamWorker;
 import org.radarcns.stream.aggregator.NumericAggregate;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -33,7 +33,7 @@ import org.slf4j.LoggerFactory;
 /**
  * Definition of Kafka Stream for aggregating temperature values collected by Empatica E4.
  */
-public class E4TemperatureStream extends StreamWorker<ObservationKey, EmpaticaE4Temperature> {
+public class E4TemperatureStream extends KStreamWorker<ObservationKey, EmpaticaE4Temperature> {
     private static final Logger logger = LoggerFactory.getLogger(E4TemperatureStream.class);
 
     public E4TemperatureStream(Collection<StreamDefinition> definitions, int numThread,
