+++ src/main/java/org/radarcns/stream/phone/PhoneUsageStream.java
@@ -22,14 +22,14 @@ import org.radarcns.kafka.ObservationKey;
 import org.radarcns.passive.phone.PhoneUsageEvent;
 import org.radarcns.stream.StreamDefinition;
 import org.radarcns.stream.StreamMaster;
-import org.radarcns.stream.StreamWorker;
+import org.radarcns.stream.KStreamWorker;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import javax.annotation.Nonnull;
 import java.util.Collection;
 
-public class PhoneUsageStream extends StreamWorker<ObservationKey, PhoneUsageEvent> {
+public class PhoneUsageStream extends KStreamWorker<ObservationKey, PhoneUsageEvent> {
     private static final Logger logger = LoggerFactory.getLogger(PhoneUsageStream.class);
 
     // 1 day until an item is refreshed
